6: Moved lambdaResponse function to utils
